Fixing bug in Update PRoduct

+++ products/product_service.py
@@ -213,27 +213,48 @@ class ProductService:
             existing_s_ids = [size.id for size in existing_product_sizes]
             existing_clr_ids = [color.id for color in existing_product_colors]
 
-            categories_changed = all(x == y for x, y in zip(existing_cat_ids, category_ids))
-            sizes_changed = all(x == y for x, y in zip(existing_s_ids, size_ids))
-            colors_changed = all(x == y for x, y in zip(existing_clr_ids, color_ids))
-            gender_changed = True if len(existing_product_gender) > 0 and \
-                                     existing_product_gender.id == gender.id else False
+            categories_changed = True if len(category_ids) != len(existing_cat_ids) or \
+                                         (len(category_ids) == len(existing_cat_ids) and sorted(category_ids) != sorted(
+                                             existing_cat_ids)) else False
+            sizes_changed = True if len(size_ids) != len(existing_s_ids) or \
+                                    (len(size_ids) == len(existing_s_ids) and sorted(size_ids) != sorted(
+                                        existing_s_ids)) else False
+            colors_changed = True if len(color_ids) != len(existing_clr_ids) or \
+                                     (len(color_ids) == len(existing_clr_ids) and sorted(color_ids) != sorted(
+                                         existing_clr_ids)) else False
+
+            gender_changed = False if len(existing_product_gender) > 0 and \
+                                      existing_product_gender[0].id == gender.id else True
+            logger.info(
+                f'---> ProductService ---> save_product_attributes ---> categories_changed: {categories_changed}')
+            logger.info(
+                f'---> ProductService ---> save_product_attributes ---> sizes_changed: {sizes_changed}')
+            logger.info(
+                f'---> ProductService ---> save_product_attributes ---> colors_changed: {colors_changed}')
+            logger.info(
+                f'---> ProductService ---> save_product_attributes ---> gender_changed: {gender_changed}')
 
             if categories_changed:
                 product.categories.clear()
 
-            if gender_changed:
+            if existing_product_gender and gender_changed:
                 repo.delete_product_attribute_value(existing_product_gender.id)
 
             if sizes_changed:
                 ids_missing_in_existing = [x for x in found_size_ids if x not in existing_s_ids]
+                logger.info(
+                    f'---> ProductService ---> save_product_attributes ---> ids_missing_in_existing: {ids_missing_in_existing}')
                 product_attribute_value_ids_to_be_deleted.extend(ids_missing_in_existing)
 
             if colors_changed:
                 ids_missing_in_existing = [x for x in found_color_ids if x not in existing_clr_ids]
+                logger.info(
+                    f'---> ProductService ---> save_product_attributes ---> ids_missing_in_existing: {ids_missing_in_existing}')
                 product_attribute_value_ids_to_be_deleted.extend(ids_missing_in_existing)
 
             if len(product_attribute_value_ids_to_be_deleted) > 0:
+                logger.info(
+                    f'---> ProductService ---> save_product_attributes ---> product_attribute_value_ids_to_be_deleted: {product_attribute_value_ids_to_be_deleted}')
                 repo.delete_product_attribute_values(product_attribute_value_ids_to_be_deleted)
 
         if not is_edit or (is_edit and categories_changed):
